src/test/java/com/myCompany/RepairAgency/servlet/request/get/realization/ShowLoginFormCommandTest.java
package com.myCompany.RepairAgency.servlet.request.get.realization;

import com.myCompany.RepairAgency.Constants;
import com.myCompany.RepairAgency.servlet.Path;
import com.myCompany.RepairAgency.servlet.PathFactory;
import com.myCompany.RepairAgency.servlet.util.AttributeFSTR;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockedStatic;
import org.mockito.Mockito;
import org.springframework.mock.web.MockHttpServletRequest;
import org.springframework.mock.web.MockHttpServletResponse;
import org.springframework.mock.web.MockHttpSession;

import java.util.stream.Collectors;
import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

class ShowLoginFormCommandTest {
    @Mock
    Path mockPath;
    @Mock
    MockHttpServletRequest request;
    @Mock
    MockHttpSession session;
    @Mock
    MockHttpServletResponse response;


    @BeforeEach
    public void initMocks() {
        response = new MockHttpServletResponse();
        mockPath = Mockito.mock(Path.class);
        request = new MockHttpServletRequest();
        session = new MockHttpSession();
        request.setSession(session);
    }

    @Test
    void execute() {
        try (MockedStatic<PathFactory> ignored2 = Mockito.mockStatic(PathFactory.class);
             MockedStatic<AttributeFSTR> ignored3 = Mockito.mockStatic(AttributeFSTR.class)) {

            Mockito.when(PathFactory.getPath(Mockito.eq("path.page.forward.guest.login"))).thenReturn(mockPath);

            assertEquals(mockPath, new ShowLoginFormCommand().execute(request, response));
            assertEquals("title.login", request.getAttribute("title"));
            ignored2.verify(() -> PathFactory.getPath("path.page.forward.guest.login"), Mockito.times(1));
        }
    }

    @Test
    void allowedUserRoles() {
        assertEquals(Stream.of(Constants.ROLE.Guest).collect(Collectors.toList()),
                new ShowLoginFormCommand().allowedUserRoles());
    }
}